removing per ryan's request
In May of 2020, we released a new document: Reusing Authorizations for Cloud Products Quick Guide. This document outlines how to reuse a security package for a FedRAMP Authorized CSO. Ultimately, we did not find any critical information about reuse in the longer document that would be lost if we removed this from the website.