Delete federalist.json
Removing based on the advice of Federalist support: "Looking at your repository, it appears you are using our addition cache-control headers option with this federalist.json file.  The options in the file appear to be caching all your files for a year so the CDN in front of your site is rarely going back to your Federalist site for new updates.  The easiest solution is to delete the federalist.json file and then slowly add new cache-control rules to optimize the site's performance.  FWIW, most of our customers do not use the custom Federalist cache rules and their sites perform well."